Tell cabal to avoid reinstalls
Reinstalls break the global package cache, so tell cabal to avoid them
when it's calculating an install plan. If it does suggest a reinstall,
then we bomb out with an error.
